test: 💍 works module e2e

+++ src/work/work.service.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   ConflictException,
   Injectable,
   Logger,
@@ -6,21 +7,20 @@ import {
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { RemoveBulkRecordsDto } from 'src/common/dto/remove-bulk-records.dto';
-import { TypeFilterDate } from 'src/common/enums/TypeFilterDate';
-import { TypeFilterNumber } from 'src/common/enums/TypeFilterNumber';
+import { getComparisonOperator } from 'src/common/helpers/get-comparison-operator';
 import { organizeIDsToUpdateEntity } from 'src/common/helpers/organize-ids-to-update-entity';
+import { HandlerErrorService } from 'src/common/services/handler-error.service';
 import { monthNamesES } from 'src/common/utils/monthNamesEs';
 import { PrinterService } from 'src/printer/printer.service';
 import { DataSource, Repository } from 'typeorm';
-import { CreateWorkDto } from './dto/create-work.dto';
-import { QueryParamsWork } from './dto/query-params-work.dto';
-import { QueryTotalWorksInYearDto } from './dto/query-params-total-works-year';
-import type { UpdateWorkDto } from './dto/update-work.dto';
 import { WorkDetailsDto } from './dto/create-work-details.dto';
+import { CreateWorkDto } from './dto/create-work.dto';
+import { QueryTotalWorksInYearDto } from './dto/query-params-total-works-year';
+import { QueryParamsWork } from './dto/query-params-work.dto';
+import type { UpdateWorkDto } from './dto/update-work.dto';
 import { WorkDetails } from './entities/work-details.entity';
 import { Work } from './entities/work.entity';
 import { getWorkReport } from './reports/get-work';
-import { HandlerErrorService } from 'src/common/services/handler-error.service';
 
 @Injectable()
 export class WorkService {
@@ -52,6 +52,7 @@ export class WorkService {
       await queryRunner.manager.save(work);
 
       await queryRunner.commitTransaction();
+      return work;
     } catch (error) {
       await queryRunner.rollbackTransaction();
       this.handlerError.handle(error);
@@ -74,6 +75,7 @@ export class WorkService {
       filter_by_total = false,
       type_filter_total,
       total,
+
       employees = [],
     } = queryParams;
     const queryBuilder = this.workRepository
@@ -86,31 +88,22 @@ export class WorkService {
       .take(limit)
       .skip(offset * limit);
 
-    if (crop.length > 0) {
+    crop.length > 0 &&
       queryBuilder.andWhere('crop.id = :cropId', { cropId: crop });
-    }
 
-    if (filter_by_date) {
-      const operation =
-        TypeFilterDate.AFTER == type_filter_date
-          ? '>'
-          : TypeFilterDate.EQUAL == type_filter_date
-            ? '='
-            : '<';
-      queryBuilder.andWhere(`work.date ${operation} :date`, { date });
-    }
+    filter_by_date &&
+      queryBuilder.andWhere(
+        `work.date ${getComparisonOperator(type_filter_date)} :date`,
+        { date },
+      );
 
-    if (filter_by_total) {
-      const operation =
-        TypeFilterNumber.MAX == type_filter_total
-          ? '>'
-          : TypeFilterNumber.EQUAL == type_filter_total
-            ? '='
-            : '<';
-      queryBuilder.andWhere(`work.total ${operation} :total`, { total });
-    }
+    filter_by_total &&
+      queryBuilder.andWhere(
+        `work.total ${getComparisonOperator(type_filter_total)} :total`,
+        { total },
+      );
 
-    if (employees.length > 0) {
+    employees.length > 0 &&
       queryBuilder.andWhere((qb) => {
         const subQuery = qb
           .subQuery()
@@ -122,17 +115,21 @@ export class WorkService {
           .getQuery();
         return 'work.id IN ' + subQuery;
       });
-    }
 
     const [works, count] = await queryBuilder.getManyAndCount();
 
+    if (works.length === 0 && count > 0) {
+      throw new NotFoundException(
+        'There are no work records with the requested pagination',
+      );
+    }
+
     return {
-      rowCount: count,
-      rows: works.map((item) => ({
-        ...item,
-        crop: { id: item.crop.id, name: item.crop.name },
-      })),
-      pageCount: Math.ceil(count / limit),
+      total_row_count: count,
+      current_row_count: works.length,
+      total_page_count: Math.ceil(count / limit),
+      current_page_count: works.length > 0 ? offset + 1 : 0,
+      records: works,
     };
   }
 
@@ -140,15 +137,18 @@ export class WorkService {
     const work = await this.workRepository.findOne({
       withDeleted: true,
       where: { id },
-      relations: { crop: true, details: { employee: true } },
+      relations: {
+        crop: true,
+        details: { employee: true, payments_work: true },
+      },
     });
-    if (!work)
-      throw new NotFoundException(`Work register with id: ${id} not found`);
+    if (!work) throw new NotFoundException(`Work with id: ${id} not found`);
     return work;
   }
 
   async update(id: string, updateWorkDto: UpdateWorkDto) {
     const work = await this.findOne(id);
+
     const queryRunner = this.dataSource.createQueryRunner();
     await queryRunner.connect();
     await queryRunner.startTransaction();
@@ -156,7 +156,7 @@ export class WorkService {
       const { details, crop, ...rest } = updateWorkDto;
 
       const oldDetails: WorkDetails[] = work.details;
-      const newDetails: WorkDetailsDto[] = details;
+      const newDetails: WorkDetailsDto[] = updateWorkDto.details ?? [];
 
       const oldIDs: string[] = oldDetails.map((record) => record.id);
       const newIDs: string[] = newDetails.map((record) =>
@@ -169,15 +169,52 @@ export class WorkService {
       );
 
       for (const recordId of toDelete) {
+        const dataRecordOld = oldDetails.find(
+          (record) => record.id === recordId,
+        );
+
+        if (dataRecordOld.payment_is_pending === false) {
+          throw new BadRequestException(
+            `You cannot delete the record with id ${recordId} , it is linked to a payment record.`,
+          );
+        }
+
+        if (dataRecordOld.deletedDate !== null) {
+          throw new BadRequestException(
+            `You cannot delete the record with id ${recordId} , it is linked to other records.`,
+          );
+        }
         await queryRunner.manager.delete(WorkDetails, recordId);
       }
 
       for (const recordId of toUpdate) {
-        const work = newDetails.find((item) => {
-          return item.id === recordId;
-        });
-        const { id, ...rest } = work;
-        await queryRunner.manager.update(WorkDetails, id, rest);
+        const dataRecordNew = newDetails.find(
+          (record) => record.id === recordId,
+        );
+        const dataRecordOld = oldDetails.find(
+          (record) => record.id === recordId,
+        );
+
+        const valuesAreDifferent =
+          dataRecordNew.value_pay !== dataRecordOld.value_pay;
+
+        if (valuesAreDifferent) {
+          switch (true) {
+            case dataRecordOld.payment_is_pending === false:
+              throw new BadRequestException(
+                `You cannot update the record with id ${recordId} , it is linked to a payment record.`,
+              );
+            case dataRecordOld.deletedDate !== null:
+              throw new BadRequestException(
+                `You cannot update the record with id ${recordId} , it is linked to other records.`,
+              );
+          }
+        }
+        await queryRunner.manager.update(
+          WorkDetails,
+          { id: recordId },
+          dataRecordNew,
+        );
       }
 
       for (const recordId of toCreate) {
@@ -192,8 +229,8 @@ export class WorkService {
       }
 
       await queryRunner.manager.update(Work, id, rest);
-
       await queryRunner.commitTransaction();
+      return this.findOne(id);
     } catch (error) {
       await queryRunner.rollbackTransaction();
       this.handlerError.handle(error);
@@ -204,6 +241,8 @@ export class WorkService {
 
   async remove(id: string) {
     const work = await this.findOne(id);
+
+    console.log(work);
 
     if (work.details.some((item) => item.payments_work !== null)) {
       throw new ConflictException(
@@ -222,9 +261,19 @@ export class WorkService {
   }
 
   async removeBulk(removeBulkWorksDto: RemoveBulkRecordsDto<Work>) {
+    const success: string[] = [];
+    const failed: { id: string; error: string }[] = [];
+
     for (const { id } of removeBulkWorksDto.recordsIds) {
-      await this.remove(id);
+      try {
+        await this.remove(id);
+        success.push(id);
+      } catch (error) {
+        failed.push({ id, error: error.message });
+      }
     }
+
+    return { success, failed };
   }
 
   async exportWorkToPDF(id: string) {
@@ -233,72 +282,70 @@ export class WorkService {
     return this.printerService.createPdf({ docDefinition });
   }
 
+  async getWorkData(year: number, cropId: string, employeeId: string) {
+    const queryBuilder = this.workRepository
+      .createQueryBuilder('work')
+      .leftJoin('work.crop', 'crop')
+      .leftJoin('work.details', 'details')
+      .leftJoin('details.employee', 'employee')
+      .select([
+        'CAST(EXTRACT(MONTH FROM work.date) AS INTEGER) as month',
+        'CAST(SUM(DISTINCT work.total) AS INTEGER) as total',
+        'COUNT(work) as quantity_works',
+      ])
+      .where('EXTRACT(YEAR FROM work.date) = :year', { year })
+      .groupBy('EXTRACT(MONTH FROM work.date)')
+      .orderBy('month', 'ASC');
+
+    if (cropId) {
+      queryBuilder.andWhere('crop.id = :cropId', { cropId });
+    }
+    if (employeeId) {
+      queryBuilder.andWhere('employee.id = :employeeId', { employeeId });
+    }
+
+    const rawData = await queryBuilder.getRawMany();
+
+    const formatData = monthNamesES.map((monthName: string, index: number) => {
+      const monthNumber = index + 1;
+      const record = rawData.find((item) => {
+        return item.month === monthNumber;
+      });
+
+      if (!record) {
+        return {
+          month_name: monthName,
+          month_number: monthNumber,
+          total: 0,
+          quantity_works: 0,
+        };
+      }
+
+      delete record.month;
+
+      return {
+        ...record,
+        month_name: monthName,
+        month_number: monthNumber,
+      };
+    });
+
+    return formatData;
+  }
+
   async findTotalWorkInYear({
     year = 2025,
-    crop = '',
-    employee = '',
+    cropId = '',
+    employeeId = '',
   }: QueryTotalWorksInYearDto) {
     const previousYear = year - 1;
 
-    const getWorkData = async (
-      year: number,
-      cropId: string,
-      employeeId: string,
-    ) => {
-      const queryBuilder = this.workRepository
-        .createQueryBuilder('work')
-        .leftJoin('work.crop', 'crop')
-        .leftJoin('work.details', 'details')
-        .leftJoin('details.employee', 'employee')
-        .select([
-          'CAST(EXTRACT(MONTH FROM work.date) AS INTEGER) as month',
-          'CAST(SUM(DISTINCT work.total) AS INTEGER) as total',
-          'COUNT(work) as quantity_works',
-        ])
-        .where('EXTRACT(YEAR FROM work.date) = :year', { year })
-        .groupBy('EXTRACT(MONTH FROM work.date)')
-        .orderBy('month', 'ASC');
-
-      if (cropId) {
-        queryBuilder.andWhere('crop.id = :cropId', { cropId });
-      }
-      if (employeeId) {
-        queryBuilder.andWhere('employee.id = :employeeId', { employeeId });
-      }
-
-      const rawData = await queryBuilder.getRawMany();
-
-      const formatData = monthNamesES.map(
-        (monthName: string, index: number) => {
-          const monthNumber = index + 1;
-          const record = rawData.find((item) => {
-            return item.month === monthNumber;
-          });
-
-          if (!record) {
-            return {
-              month_name: monthName,
-              month_number: monthNumber,
-              total: 0,
-              quantity_works: 0,
-            };
-          }
-
-          delete record.month;
-
-          return {
-            ...record,
-            month_name: monthName,
-            month_number: monthNumber,
-          };
-        },
-      );
-
-      return formatData;
-    };
-
-    const currentYearData = await getWorkData(year, crop, employee);
-    const previousYearData = await getWorkData(previousYear, crop, employee);
+    const currentYearData = await this.getWorkData(year, cropId, employeeId);
+    const previousYearData = await this.getWorkData(
+      previousYear,
+      cropId,
+      employeeId,
+    );
 
     const workDataByYear = [
       { year, data: currentYearData },
